[r=julian-edwards][bug=1044206][author=jtv] Create host maps for new DHCP leases, using the database as the store of already-known leases.
This was a missing link: we weren't yet computing which leases were new and triggering their registration as host maps.

Abandons the new_leases parameter to the update_leases API call; it's now the server that figures out which leases are new, based on database state.  This does mean that the server relies on the task for creating new host maps being retried on failure.  As far as the server is concerned, the new host maps are sure to be registered as soon as the task to do so goes out.