[r=julian-edwards][bug=1044203][author=jtv] Stop updating DNS zone files when there are no new DHCP leases.
As discussed with Gavin, only additions really matter.  If there are none of those, there's no need to rewrite zone files.  Disappearing leases need no immediate DNS updates (and see bug 1045277).

The “post_updates” signal was a bit generic to skip in its entirety with such intricate justification, but this was its only use and it only had one handler.  After deliberation with Raphaël I replaced it with a conventional function call so that the action I'm skipping is clear and explicit in the calling code.

Also after deliberation with Raphaël, I restricted the zone-file update to the nodegroup that has changed.  We don't need to rewrite all DNS zones just because one had a change.