[r=julian-edwards][bug=][author=jtv] Rename NodeGroup.dhcp_interface to dhcp_interfaces.
This field is really supposed to be a space-separated list of interfaces.  I think a maximum length of 255 characters is still good enough though.

South originally generated a migration that removed one column and added another, but I edited it manually to rename the existing column.  And vice versa for the reverse migration.