[r=rvb][bug=1047061][author=jtv] Set DHCP next_server option to point nodes to the TFTP server.
This is how the nodes know where to download their kernels and initrds.  Once we have distributed TFTP, the current setting of self.worker_ip will be correct again but for now, it needs to be the MAAS-facing address of the central server.

Once we do start using self.worker_ip again, we will have to make sure we get the right address for the master nodegroup.  Its IP worker_ip is currently still hardcoded to 127.0.0.1.  When we change that, we will have to pay attention to bootstrapping: the administrator may not have had a chance to change the DEFAULT_MAAS_URL setting yet on which we would base the master's IP address.