[r=julian-edwards][bug=1025582][author=jtv] New API call for the master worker: register_boot_images.
This is how the master worker will report to the MAAS server which images it has available for net-booting nodes.  The API code stores the results in BootImage (see predecessor branch) so that the various operations for booting nodes can check that the required images are really available.  If not, it can report the problem to the user and make recommendations for running or reconfiguring the maas-import-pxe-files script.

The new call lives on a new API path, boot-images/.  I found it didn't quite fit in anywhere.  It might have fit at the root, but we're not making a habit of that.  Plus, we may want to add querying abilities to the boot images later, for which this separate path is just right.

Images are passed in as a JSON-encoded list of dicts.  I nearly used a list of tuples, but that would have been error-prone and harder to keep compatible between API versions.  The calls aren't very data-intensive so there's little reason to keep the encoding minimal.

This branch is to be backported to 1.1.