[r=julian-edwards][bug=1025582][author=jtv] Have the master worker report its available boot images.
This creates a new task for the master worker, and runs it on a 5-minute schedule.  The task scans the TFTP filesystem tree (which is on the same system as the master worker) for boot images, and reports what it finds to the API.  In a subsequent branch, the UI will report missing images as a persistent error.