[r=allenap,smoser][bug=][author=racb] Add an armhf install template variant
This is the same as the cross-arch install template, except that it doesn't use the SAY command which doesn't exist in U-Boot at the moment.

Also explicitly rename the existing install pxelinux.cfg template to make it clearer that it is not used for any other boot purpose. Since we already have a config.commissioning.template and a config.local.template, the only use of this file is as an install template. By explicitly naming it, I can override it for armhf, which is something that I can't do for the final fallback filename of config.template.

Also alter some tests that assumed that a default template exists, since it now there isn't one. Add a conditional skip on the default template test if the default template does not exist, and switch some tests that used the default template and assumed it looked like an install template to explicitly request an install template.