[r=rvb,jameinel][bug=][author=jtv] Allow node-group selection when creating a node in API or UI.
Node groups are selected by subnet: any IP address in a node group's subnet identifies that node group.  It could be the node's IP, or the cluster manager's, or the broadcast address, or any other IP in the subnet.  This requires a node group to have an interface (subnet) defined before nodes can be added to it, and an IP in the subnet must not be in any other node group's subnet.