[r=jameinel][bug=1049933][author=allenap] Defines a new operations dispatch mechanism on top of Piston.
The existing class decoration code was confusing. This is still somewhat complex, but much less so, and the flow of control is easier to trace. This also takes over the definition of allowed_methods: all operations are exported, and CRUD methods that have not been erased (set to None) are also exported.