[r=julian-edwards][bug=][author=julian-edwards] specify 'preserve_sources_list' to cloud-init in install preseed
This moves the cloud-init/local-cloud-config out of the preseed explicitly
and instead renders it in compose_cloud_init_preseed.

We now put 2 items in this debconf escaped yaml, essentially feeding
cloud-init this yaml:
  manage_etc_hosts: localhost
  preserve_source_list: true

preserve_sources_list is required so cloud-init on first boot does not:
 a.) break arm mirrors (12.04 arm installs)
 b.) override a local mirror that was specified somewhere else.

In the future this will make adding other cloud-init config simpler.

Also, remove incorrect and useless [ignored] string in commissioning user-data.