[r=flacoste][bug=1065499][author=smoser] if ipmi-locate fails, but /dev/ipmi[0-9] exists continue down ipmi path
on quantal, the version of ipmi-locate in freeipmi (1.1.5-3ubuntu2) was
failing to locate ipmi on armhf boards. ipmi-locate did produce a postive
result on precise (0.8.12-3ubuntu1).

This works around that, and continues down the path if the /dev/ipmi
device is present.

Also, added here is a 'udevadm settle', which is really required any time
you load a module to prevent a race condition before attempting to access
the named device.

It would seem likelyhood of false positive here is very small, and at
this point it would seem that the end result would be a node enlisted
with empty power settings.