[r=allenap][bug=][author=jameinel] Change the MAAS application so that it can return content with 'Content-Encoding: gzip'
This should be useful for general web browsers, so that any large page gets compressed before being sent on the wire. It helps the Tag builders immensely, because we are requesting many MB of data in each request, which compresses very well. This also teaches the MAASDispatcher how to set the Accept-Encoding header, and how to handle the Content-Encoding response.

In testing with many nodes, rebuilding tags can saturate a 1GB network link. So compressing the stream speeds up the tag processing from 22s to 15s on the nodes that have compression enabled.