[r=allenap,rvb,andreserl][bug=1075313][author=smoser] in ephemeral/commissioning boot, specify root= by iscsi device path
previously we were booting iscsi root with 'LABEL=cloudimg-rootfs'.
There were a few issues with that:
 a.) that implied/required that all ephemeral images would have this root
     filesystem label.
 b.) if there was a filesystem on the local system with the given label
     the selected device would be non-determinable.
     This could happen innocently, or potentially on purpose as an attempt
     to retain control of a system after it had been returned.

By booting with a /dev/disks/by-path, we're explicitly telling the
initramfs that is being sent that we want the root device to come over
iscsi.  The tradeoff here is that we're now expecting the kernel/initramfs
to consistently provide this path.