[r=rvb][bug=][author=jtv] Show network connections on Node detail page.
Network connections are listed with their MAC addresses.  So a MACs display of “11:11:11:11:11:11 (on management, staging); 22:22:22:22:22:22; 33:33:33:33:33:33 (on management)” means that the node has 3 known NICs: the NIC with MAC 11<etc> is connected to the ‘management’ and ‘staging’ networks (in practice, at least one of those would have to be a VLAN), the NIC with MAC 22<etc> is not known to be connected to any networks, and the NIC with MAC 33<etc> is known to be connected only to the ‘management’ network.

The HTML templating for this is ugly.  Can't be reformatted, or we'll get whitespace in unwanted places.