[r=gmb][bug=1339631][author=jtv] Fix find_nodegroups for IPv6.  Trying to ‘AND’ a remote IPv6 address taken from an API request with an IPv4 cluster interface netmask crashed.
It's tempting just to add a simple check that the addresses being compared must be in the same address family.  But SQL does not guarantee short-circuit evaluation: it may choose to evaluate the ‘AND’ before checking address families, or always evaluate both conditions.  The BETWEEN operator does what we need: the request address must lie between (inclusive) the cluster interface address with all host bits cleared, and the cluster interface address with all host bits set.