[r=gmb][bug=][author=jtv] Add NodeGroupInterface.name field, and during upgrade, populate it based on the existing interface field.
This branch is not meant for landing in isolation: a cluster interface is now identified by a cluster and a name, rather than a cluster and an interface.  The API still needs to switch to the new key, but this branch is big enough as it is.  (The DHCP code will continue to use the interface field.)  The field can be left blank on input, but the code ensures that it is never blank in the database.

The changes are made in a single schema migration, because migrations are relatively expensive.  The migration will use the existing interface name unchanged where it can, for compatibility, but make up unique names for unnamed interfaces and perform very basic escaping on weird characters.  The interface field never had a validator, so there could in principle be wildly unexpected characters in there.  I intend to add a branch to accept the colon as used in ethernet aliases (“eth0:1”) unchanged, for increased backward compatibility.