[r=gmb][bug=][author=jtv] Identify the cluster interfaces to which a MACAddress is attached, including both IPv4 and IPv6, so that next we can assign static IPv6 addresses in addition to IPv4 ones.
We should probably replace MACAddress.cluster_interface with an m:n relationship, but for now, Andres came up with a simpler scheme: look for cluster interfaces that use the same network interface on the cluster controller as the cluster interface that manages the MAC's IPv4 address.  Encapsulating this in a method should reduce the impact of a future schema change.