[r=rvb][bug=][author=allenap] Retry one more time at the end.
The `retries` function in provisioningserver.utils.twisted was suggesting a final wait of up to `interval` seconds, but then not allowing the user to retry their operation (in typical use). It now always allows a final retry.

The `retries` function in maastesting.utlls suffered a similar problem. It was not used, so has been removed.