[r=julian-edwards][bug=1361967][author=allenap] Keeps the node power monitoring loop going even when there are transient failures.
Previously an unhandled error in the monitor would cause the service to halt. Now the error is logged instead.

This also adds tests for NodePowerMonitorService which did not exist before. It also introduces a new NoSuchCluster exception which is raised when the region has no knowledge of the calling cluster.