[r=blake-rouse][bug=][author=rvb] Rename NODE_STATUS.FAILED_TESTS into NODE_STATUS.FAILED_COMMISSIONING.
As soon as the new twisted-based PowerOn/PowerOff methods will be wired up, many things will be able to transition a commissioning node into a broken state; i.e. not only the failure of the commissioning tests themselves but many kinds of problems like failure to power on, etc.

Also, this renaming is in sync with the introduction of the new FAILED_DEPLOYMENT state: each 'in progress' state gets a corresponding 'failed' state.