[r=rvb][bug=1363900][author=jtv] When running a dev environment, move hard-coded "absolute" paths into the run/ directory.
This branch introduces a new way to compose such paths, which supports injection of a faux root directory through the MAAS_ROOT environment variable.  It unbreaks startup of the dev server.  To avoid import side effects, the setting is now wrapped in a getter function.  We would like to do the migration of more of the hard-coded paths from celery configs in this way.