[r=gmb][bug=][author=jtv] Make disambiguation of IPv4/IPv6 cluster interface naming more user-friendly.
When multiple cluster interfaces use the same network interface on the same cluster controller (and they are not given explicit names), they get unique names based on the network interface.  If the network interface is eth9, then the first cluster interface will be eth9, and the other ones will have different names like eth9-*.  In this branch, that eth9-* is made more user-friendly: eth9-ipv4, eth9-ipv6-*.

Also, in the typical case of one IPv4 cluster interface and one IPv6 cluster interface sharing a network interface, the IPv4 will be the one that gets the same name as the network interface without suffix, e.g. ‘eth9’.