[r=gmb][bug=1365850][author=jtv] Fix new DHCP probing service: it was using cluster interface names as network interface names.
The change turned out to be more pervasive than I expected: the cluster controller calls the region in order to obtain a list of cluster interfaces, and that list didn't include the actual network interfaces to probe!  The implementation of that call wasn't unit-tested, so I had to add tests for that as well.

When the cluster controller reports foreign DHCP servers to the region controller, it does still pass the cluster interface name, not the network interface.  I added a bit of docstring to make this explicit.