[r=blake-rouse,allenap][bug=1366108][author=blake-rouse] Use get_boot_images_for RPC call in pxeconfig API call. This removes the need for BootImage table and contacts the cluster directly.
This exposed an issue in the WindowsBootMethod. It used a blocking call to the pxeconfig API, which blocked the reactor from running in pserv. This caused a race condition where the region wants to talk to the cluster, while the cluster is waiting for a response from the region... ;-) Modified the TFTPService to allow boot methods to return a deferred and fixed WindowBootMethod to use this new ability.