[r=rvb][bug=][author=jtv] Configure a node's static IPv6 addresses during Curtin installation.
Use Curtin preseeds to upload the configure_interfaces.py script onto an installing node, and to run it on the installed system's /etc/network config (if present).  If the script finds an OS installed on the filesystem that it doesn't know how to configure, it logs a warning but exits successfully so as not to break installation.