[r=rvb][bug=][author=blake-rouse] Add other operating systems and generated images to the images UI.
Adds two new section to the images UI. The sections are only shown if the boot sources provide other operating systems besides Ubuntu, and generate only shows up once a generated image is uploaded. Once MAAS supports the ability to generate images, a generate button will be added to the generated images section and it will always be shown.