[r=gmb][bug=][author=jtv] Select network interfaces on a node that should be brought up automatically on boot.  This is part of the work for writing its /etc/network/interfaces (assuming it's installed with a Debian-like OS).  That code is not active yet, but is now very close to completion (and has been tested in a spike).
We chose the selection behaviour as being the closest to existing behaviour, not as the ideal behaviour as we would without the time constraints.  In the normal situation where a node has one network interface on a network where MAAS manages DHCP, it is identical to what happens in the current code.  If the node has multiple network interfaces on one or more MAAS-managed networks, they will all come up; the existing code will only bring up the interface from which the node netbooted, and some people seem to run into problems in this situation.  And finally, if the node is not connected to any MAAS-managed DHCP, we have no knowledge of how the machine netbooted, so we bring up all interfaces.