[r=jtv][bug=1330765][author=gmb] Add a start() method to the Node model.
This method is largely a distillation of what happens in bulk in NodeManager.start_nodes; most of the tests have been adapted from the tests for that method.

This is the first of several branches to move us away from batching starting and stopping nodes; Node.start() is not currently used anywhere and won't be until the other work is complete.